requirements.txt
flask==3.0.3
werkzeug==3.0.3
python-dotenv==1.0.1
gunicorn==22.0.0
